pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.4",
    "pandas>=2.2.3",
    "plotly>=6.0.0",
    "streamlit>=1.42.2",
    "yfinance>=0.2.54",
]
